Move AnsibleAWSModule specifc code out into modules.py
